ppx-ocr/pyproject.toml
[project]
name = "ppx-ocr"
version = "0.1.0"
description = "OCR pipeline powered by PaddleOCR"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=14.0",
    "pymupdf>=1.26.7",
    "joblib>=1.5.3",
    "pandas>=2.0",
    "fastparquet>=2025.12.0",
    "pandera>=0.29.0",
    "pillow>=10.0",
    # paddlepaddle-gpu, paddleocr, paddlex installed separately via Makefile
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ppx-ocr = "ppx_ocr.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ppx_ocr"]
